pyproject.toml
[tool.poetry]
name = "CatsAndChocolateServer"
version = "0.1.0"
description = ""
authors = ["toru.yamada@excite-software.co.jp"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.8"
djangorestframework = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"